+++ common/src/main/java/net/mehvahdjukaar/amendments/common/block/LiquidCauldronBlock.java
@@ -1,5 +1,7 @@
 package net.mehvahdjukaar.amendments.common.block;
 
+import net.mehvahdjukaar.amendments.common.network.ModNetwork;
+import net.mehvahdjukaar.amendments.common.network.PlaySplashParticlesPacket;
 import net.mehvahdjukaar.amendments.common.recipe.RecipeUtils;
 import net.mehvahdjukaar.amendments.common.tile.LiquidCauldronBlockTile;
 import net.mehvahdjukaar.amendments.configs.CommonConfigs;
@@ -29,8 +31,10 @@ import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectInstance;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
+import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.Items;
 import net.minecraft.world.item.alchemy.PotionUtils;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.LevelAccessor;
@@ -149,18 +153,34 @@ public class LiquidCauldronBlock extends ModCauldronBlock {
         if (state.getValue(BOILING) && entity instanceof LivingEntity) {
             entity.hurt(new DamageSource(ModRegistry.BOILING_DAMAGE.getHolder()), 1.0F);
         }
-        if (entity instanceof LivingEntity living && entity.mayInteract(level, pos) &&
-                level.getBlockEntity(pos) instanceof LiquidCauldronBlockTile tile) {
+        if (entity.mayInteract(level, pos) && level.getBlockEntity(pos) instanceof LiquidCauldronBlockTile tile) {
 
             SoftFluidStack fluid = tile.getSoftFluidTank().getFluid();
             PotionNBTHelper.Type potType = getPotType(fluid);
-            if (potType != null && potType != PotionNBTHelper.Type.REGULAR &&
-                    applyPotionFluidEffects(level, pos, living, fluid)) {
-                tile.consumeOneLayer();
+            if (entity instanceof LivingEntity living) {
+                if (potType != null && potType != PotionNBTHelper.Type.REGULAR &&
+                        applyPotionFluidEffects(level, pos, living, fluid)) {
+                    tile.consumeOneLayer();
+                }
+
+                if (CompatHandler.ALEX_CAVES) {
+                    AlexCavesCompat.acidDamage(fluid, level, pos, state, entity);
+                }
             }
-            if (CompatHandler.ALEX_CAVES) {
-                AlexCavesCompat.acidDamage(fluid, level, pos, state, entity);
+
+            if (!tile.isGlowing() && potType != null && entity instanceof ItemEntity ie
+                    && ie.getItem().is(Items.GLOW_INK_SAC)
+                    && isEntityInsideContent(state, pos, entity)
+            ) {
+                ModCauldronBlock.playSplashEffects(entity, getContentHeight(state));
+
+                tile.setGlowing(true);
+                ie.getItem().shrink(1);
+                if (ie.getItem().isEmpty()) {
+                    ie.discard();
+                }
             }
+
         }
     }