+++ common/src/main/java/net/mehvahdjukaar/amendments/common/network/ModNetwork.java
@@ -7,10 +7,10 @@ import net.mehvahdjukaar.moonlight.api.platform.network.NetworkDir;
 public class ModNetwork {
 
     public static void init() {
-
     }
 
     public static final ChannelHandler CHANNEL = ChannelHandler.builder(Amendments.MOD_ID)
             .register(NetworkDir.PLAY_TO_SERVER, SyncLecternBookMessage.class, SyncLecternBookMessage::new)
+            .register(NetworkDir.PLAY_TO_CLIENT, PlaySplashParticlesPacket.class, PlaySplashParticlesPacket::new)
             .build();
 }